SlicePredicate: Added support for "default" ranges
When using SlicePredicate for range queries, Hector expects both start and
finish values to be set. This is not the best behavior.

For example, Cassandra allows to query for the "first N columns" simply by
setting start and finish values to empty string and setting count to N.
Although this query can be performed with String/StringSerializer, with
UUID/UUIDSerializer it can't be done. To get "first N columns" of type UUID,
one has to generate min/max UUID's and use these as range values.